2022/03/17

+++ views/js/customer/order.js
@@ -13,69 +13,72 @@ function tableItemsRemomver() {
   // console.log(trash_alt);
   for (let i = 0; i < trash_alt.length; i++) {
     trash_alt[i].addEventListener("click", () => {
-      cartIdToRemove =
-        trash_alt[i].parentElement.parentElement.parentElement.childNodes[1]
-          .childNodes[1].value;
-      cartIdToRemove = parseInt(cartIdToRemove);
-      //console.log("deleting....", i, " id ", cartIdToRemove);
+      if (confirm("Are you sure you want to remove the item from your cart")) {
+        cartIdToRemove =
+          trash_alt[i].parentElement.parentElement.parentElement.childNodes[1]
+            .childNodes[1].value;
+        cartIdToRemove = parseInt(cartIdToRemove);
+        // console.log("deleting....", i, " id ", cartIdToRemove);
 
-      for (var j in productInCart) {
-        if (productInCart[j].id == cartIdToRemove) {
-          noOfItems = parseInt(noOfItems);
-          noOfItems = noOfItems - productInCart[j].InCart;
-          localStorage.setItem("cartNumbers", noOfItems);
-          document.querySelector(".fa-shopping-cart-span").textContent =
-            noOfItems;
-          // console.log(noOfItems, cartCost);
-
-          cartCost = parseInt(cartCost);
-          cartCost -= productInCart[j].InCart * productInCart[j].price;
-          if (cartCost < 1) {
-            cartCost = 0;
-          }
-          localStorage.setItem("totalCost", cartCost);
-          total.innerHTML = `${parseInt(cartCost)}`;
-          // console.log(cartCost);
-
-          // buy_all button render start
-          buy_all = document.querySelector("#buy_all");
-          cartCout = Object.keys(productInCart).length;
-          cartCout -= 1;
-          console.log(cartCout, buy_all);
-          if (cartCout > 1) {
-            buy_all.innerHTML =
-              "<a href='Payment/RenderBuyAll/" +
-              cartCout +
-              "' class='btn'>Buy All</a>";
-            console.log(cartCout, buy_all);
-          }
-          //  buy_all button render end
-
-          // console.log("items id: ", productInCart[j].id, i, "in cart :", productInCart[j].InCart, " cartId: ", productInCart[j].cartId);
+        for (var j in productInCart) {
           if (productInCart[j].id == cartIdToRemove) {
-            // console.log("deletetion process...:", productInCart[j].cartId);
-            delete productInCart[j];
-            localStorage.setItem(
-              "productInCart",
-              JSON.stringify(productInCart)
-            );
-            trash_alt[
-              i
-            ].parentElement.parentElement.parentElement.style.display = "none";
-            cart_items = document.querySelectorAll(".cart-item");
-            if (cart_items.length > 0) {
-              // console.log(cart_items);
-              for (let k = 0; k < cart_items.length; k++) {
-                if (
-                  cart_items[k].childNodes[1].childNodes[1].value ==
-                  cartIdToRemove
-                ) {
-                  cart_items[k].style.display = "none";
+            noOfItems = parseInt(noOfItems);
+            noOfItems = noOfItems - productInCart[j].InCart;
+            localStorage.setItem("cartNumbers", noOfItems);
+            document.querySelector(".fa-shopping-cart-span").textContent =
+              noOfItems;
+            // console.log(noOfItems, cartCost);
+
+            cartCost = parseFloat(cartCost);
+            cartCost -= productInCart[j].InCart * productInCart[j].price;
+            if (cartCost < 1) {
+              cartCost = 0;
+            }
+            localStorage.setItem("totalCost", cartCost);
+            total.innerHTML = `${parseFloat(cartCost).toFixed(2)}`;
+            // console.log(cartCost);
+
+            // buy_all button render start
+            buy_all = document.querySelector("#buy_all");
+            cartCout = Object.keys(productInCart).length;
+            cartCout -= 1;
+            // console.log(cartCout, buy_all);
+            if (cartCout > 1) {
+              buy_all.innerHTML =
+                "<a href='Payment/RenderBuyAll/" +
+                cartCout +
+                "' class='btn'>Buy All</a>";
+              // console.log(cartCout, buy_all);
+            }
+            //  buy_all button render end
+
+            // console.log("items id: ", productInCart[j].id, i, "in cart :", productInCart[j].InCart, " cartId: ", productInCart[j].cartId);
+            if (productInCart[j].id == cartIdToRemove) {
+              // console.log("deletetion process...:", productInCart[j].cartId);
+              delete productInCart[j];
+              localStorage.setItem(
+                "productInCart",
+                JSON.stringify(productInCart)
+              );
+              trash_alt[
+                i
+              ].parentElement.parentElement.parentElement.style.display =
+                "none";
+              cart_items = document.querySelectorAll(".cart-item");
+              if (cart_items.length > 0) {
+                // console.log(cart_items);
+                for (let k = 0; k < cart_items.length; k++) {
+                  if (
+                    cart_items[k].childNodes[1].childNodes[1].value ==
+                    cartIdToRemove
+                  ) {
+                    cart_items[k].style.display = "none";
+                  }
                 }
               }
+              // console.log("deletetion complete...");
+              DeleteUpdate(cartIdToRemove);
             }
-            // console.log("deletetion complete...");
-            DeleteUpdate(cartIdToRemove);
           }
         }
       }
@@ -130,7 +133,7 @@ function displayCheckout() {
     total.innerHTML = `${parseFloat(totalCost).toFixed(2)}`;
 
     buy_all = document.querySelector("#buy_all");
-    console.log(Object.keys(cartItems).length, buy_all);
+    // console.log(Object.keys(cartItems).length, buy_all);
 
     buy_all.innerHTML =
       `<form action='Payment/RenderBuyAll/` +
@@ -152,43 +155,39 @@ function displayCheckout() {
 
 // displayCheckout();
 function item_remover(id) {
-  // console.log("hello", id);
-  let trash_alt = document.querySelectorAll(".fa-trash-alt");
-  // console.log(trash_alt);
-  for (let k = 0; k < trash_alt.length; k++) {
-    if (
-      trash_alt[k].parentElement.parentElement.parentElement.childNodes[1]
-        .childNodes[1].value == id
-    ) {
-      // send request to remove the items from here
-      var itemToRemove = new FormData();
-      itemToRemove.append(
-        "itemToRemove",
-        trash_alt[k].parentElement.parentElement.parentElement.childNodes[3]
-          .childNodes[1].value
-      );
-      var xhr = new XMLHttpRequest();
-      xhr.open("POST", "Order/Remove");
-      xhr.send(itemToRemove);
-      // console.log(
-      //   typeof parseInt(
-      //     trash_alt[k].parentElement.parentElement.parentElement.childNodes[9]
-      //       .childNodes[1].childNodes[1].textContent
-      //   )
-      // );
-      trash_alt[k].parentElement.parentElement.parentElement.style.display =
-        "none";
-      let total = document.querySelector("#stotal");
-      // console.log(total.textContent);
-      Total = parseInt(total.textContent);
-      Total =
-        Total -
-        parseInt(
-          trash_alt[k].parentElement.parentElement.parentElement.childNodes[9]
-            .childNodes[1].childNodes[1].textContent
+  if (confirm("Are sure you want to remove the item")) {
+    let trash_alt = document.querySelectorAll(".fa-trash-alt");
+    // console.log(trash_alt);
+    for (let k = 0; k < trash_alt.length; k++) {
+      if (
+        trash_alt[k].parentElement.parentElement.parentElement.childNodes[1]
+          .childNodes[1].value == id
+      ) {
+        // send request to remove the items from here
+        var itemToRemove = new FormData();
+        itemToRemove.append(
+          "itemToRemove",
+          trash_alt[k].parentElement.parentElement.parentElement.childNodes[3]
+            .childNodes[1].value
         );
-      total.innerHTML = `${parseInt(Total)}`;
-      DeleteUpdate(id);
+        var xhr = new XMLHttpRequest();
+        xhr.open("POST", "Order/Remove");
+        xhr.send(itemToRemove);
+
+        trash_alt[k].parentElement.parentElement.parentElement.style.display =
+          "none";
+        let total = document.querySelector("#stotal");
+        // console.log(total.textContent);
+        Total = parseFloat(total.textContent);
+        Total =
+          Total -
+          parseFloat(
+            trash_alt[k].parentElement.parentElement.parentElement
+              .childNodes[11].childNodes[1].childNodes[1].textContent
+          );
+        total.innerHTML = `${parseFloat(Total).toFixed(2)}`;
+        DeleteUpdate(id);
+      }
     }
   }
 }
